requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#


about-time==3.1.1
    # via alive-progress
accumulation-tree==0.6.2
    # via tdigest
aiodns==3.0.0
    # via mercury-http
aiokafka==0.7.2
    # via -r requirements.in
aiopg==1.3.3
    # via -r requirements.in
aioredis==2.0.1
    # via -r requirements.in
alive-progress==2.4.1
    # via -r requirements.in
anyio==3.5.0
    # via
    #   starlette
    #   watchgod
art==5.5
    # via -r requirements.in
asgiref==3.5.0
    # via uvicorn
asn1crypto==1.5.1
    # via
    #   oscrypto
    #   snowflake-connector-python
async-timeout==4.0.2
    # via
    #   aiopg
    #   aioredis
    #   redis
asyncio-pool==0.5.2
    # via -r requirements.in
boto3==1.21.42
    # via -r requirements.in
botocore==1.24.42
    # via
    #   boto3
    #   s3transfer
broadkast==0.1.6
    # via -r requirements.in
cachetools==5.0.0
    # via google-auth
cassandra-driver==3.25.0
    # via -r requirements.in
certifi==2021.10.8
    # via
    #   kubernetes
    #   requests
    #   snowflake-connector-python
cffi==1.15.0
    # via
    #   cryptography
    #   pycares
    #   snowflake-connector-python
charset-normalizer==2.0.12
    # via
    #   requests
    #   snowflake-connector-python
click==8.1.2
    # via
    #   flask
    #   geomet
    #   uvicorn
configargparse==1.5.3
    # via py3cli-tools
cryptography==36.0.2
    # via
    #   pyopenssl
    #   snowflake-connector-python
datadog==0.42.0
    # via -r requirements.in
dateparser==1.1.1
    # via prometheus-api-client
deprecated==1.2.13
    # via redis
dill==0.3.4
    # via -r requirements.in
dnspython==2.2.1
    # via
    #   email-validator
    #   eventlet
easy-logger-py==0.1.2
    # via
    #   -r requirements.in
    #   broadkast
    #   py3cli-tools
    #   statserve
    #   statstream-py
email-validator==1.1.3
    # via fastapi
eventlet==0.33.0
    # via -r requirements.in
fastapi[all]==0.75.2
    # via broadkast
flask==2.1.1
    # via -r requirements.in
geomet==0.2.1.post1
    # via cassandra-driver
google-api-core==2.7.2
    # via
    #   google-cloud-core
    #   google-cloud-storage
google-auth==2.6.5
    # via
    #   google-api-core
    #   google-cloud-core
    #   google-cloud-storage
    #   kubernetes
google-cloud-core==2.3.0
    # via google-cloud-storage
google-cloud-storage==2.3.0
    # via -r requirements.in
google-crc32c==1.3.0
    # via google-resumable-media
google-resumable-media==2.3.2
    # via google-cloud-storage
googleapis-common-protos==1.56.0
    # via google-api-core
grapheme==0.6.0
    # via alive-progress
graphql-core==3.2.1
    # via mercury-http
greenlet==1.1.2
    # via
    #   eventlet
    #   playwright
grpcio==1.44.0
    # via
    #   broadkast
    #   grpcio-tools
    #   statserve
grpcio-tools==1.44.0
    # via
    #   broadkast
    #   statserve
h11==0.13.0
    # via uvicorn
h2==4.1.0
    # via mercury-http
hpack==4.0.0
    # via h2
httptools==0.4.0
    # via uvicorn
hyperframe==6.0.1
    # via h2
idna==2.9
    # via
    #   anyio
    #   broadkast
    #   email-validator
    #   requests
    #   snowflake-connector-python
    #   statserve
importlib-metadata==4.11.3
    # via flask
itsdangerous==2.1.2
    # via
    #   fastapi
    #   flask
jinja2==3.1.1
    # via
    #   fastapi
    #   flask
jmespath==1.0.0
    # via
    #   boto3
    #   botocore
joblib==1.1.0
    # via scikit-learn
kafka-python==2.0.2
    # via aiokafka
kubernetes==23.3.0
    # via broadkast
markupsafe==2.1.1
    # via jinja2
mercury-http==0.4.2
    # via -r requirements.in
motor==2.5.1
    # via -r requirements.in
numpy==1.22.3
    # via
    #   pandas
    #   prometheus-api-client
    #   scikit-learn
    #   scipy
oauthlib==3.2.0
    # via requests-oauthlib
orjson==3.6.8
    # via fastapi
oscrypto==1.3.0
    # via snowflake-connector-python
packaging==21.3
    # via
    #   broadkast
    #   redis
    #   statserve
pandas==1.4.2
    # via prometheus-api-client
playwright==1.21.0
    # via -r requirements.in
prometheus-api-client==0.5.0
    # via -r requirements.in
prometheus-client==0.14.1
    # via -r requirements.in
protobuf==3.20.0
    # via
    #   google-api-core
    #   google-cloud-storage
    #   googleapis-common-protos
    #   grpcio-tools
psutil==5.9.0
    # via -r requirements.in
psycopg2-binary==2.9.3
    # via
    #   -r requirements.in
    #   aiopg
py3-async-tools==0.1.0
    # via
    #   -r requirements.in
    #   mercury-http
py3cli-tools==0.1.4
    # via
    #   -r requirements.in
    #   broadkast
    #   statserve
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pycares==4.1.2
    # via aiodns
pycparser==2.21
    # via cffi
pycryptodomex==3.14.1
    # via snowflake-connector-python
pydantic==1.9.0
    # via fastapi
pyee==8.1.0
    # via playwright
pyjwt==2.3.0
    # via snowflake-connector-python
pymongo==3.12.3
    # via motor
pyopenssl==21.0.0
    # via snowflake-connector-python
pyparsing==3.0.8
    # via packaging
python-dateutil==2.8.2
    # via
    #   botocore
    #   dateparser
    #   kubernetes
    #   pandas
python-dotenv==0.20.0
    # via uvicorn
python-multipart==0.0.5
    # via fastapi
pytz==2022.1
    # via
    #   -r requirements.in
    #   dateparser
    #   pandas
    #   snowflake-connector-python
pytz-deprecation-shim==0.1.0.post0
    # via tzlocal
pyudorandom==1.0.0
    # via tdigest
pyyaml==6.0
    # via
    #   fastapi
    #   kubernetes
    #   uvicorn
redis==4.2.2
    # via -r requirements.in
regex==2022.3.2
    # via dateparser
requests==2.27.1
    # via
    #   datadog
    #   fastapi
    #   google-api-core
    #   google-cloud-storage
    #   kubernetes
    #   prometheus-api-client
    #   requests-oauthlib
    #   snowflake-connector-python
requests-oauthlib==1.3.1
    # via kubernetes
rsa==4.8
    # via google-auth
s3transfer==0.5.2
    # via boto3
scikit-learn==1.0.2
    # via -r requirements.in
scipy==1.8.0
    # via
    #   -r requirements.in
    #   scikit-learn
six==1.16.0
    # via
    #   cassandra-driver
    #   eventlet
    #   geomet
    #   google-auth
    #   grpcio
    #   kubernetes
    #   pyopenssl
    #   python-dateutil
    #   python-multipart
sniffio==1.2.0
    # via anyio
snowflake-connector-python==2.7.6
    # via -r requirements.in
starlette==0.17.1
    # via fastapi
statserve==0.1.5
    # via -r requirements.in
statstream-py==0.1.6
    # via
    #   -r requirements.in
    #   statserve
stringcase==1.2.0
    # via
    #   broadkast
    #   statserve
tdigest==0.5.2.2
    # via statstream-py
threadpoolctl==3.1.0
    # via scikit-learn
typing-extensions==4.2.0
    # via
    #   aioredis
    #   pydantic
tzdata==2022.1
    # via pytz-deprecation-shim
tzlocal==4.2
    # via
    #   -r requirements.in
    #   dateparser
ujson==5.2.0
    # via fastapi
urllib3==1.26.9
    # via
    #   botocore
    #   kubernetes
    #   requests
uvicorn[standard]==0.17.6
    # via fastapi
uvloop==0.16.0
    # via
    #   -r requirements.in
    #   uvicorn
uwsgi==2.0.20
    # via -r requirements.in
watchgod==0.8.2
    # via uvicorn
websocket-client==1.3.2
    # via kubernetes
websockets==10.1
    # via
    #   playwright
    #   uvicorn
werkzeug==2.1.1
    # via flask
wrapt==1.14.0
    # via deprecated
zipp==3.8.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
